move `walk` utils in their own file
